requirements.txt
astroid==1.5.3
cachetools==2.0.0
certifi==2017.4.17
chardet==3.0.4
dill==0.2.7.1
future==0.16.0
gapic-google-cloud-language-v1beta2==0.15.3
google-auth==1.0.2
google-auth-httplib2==0.0.2
google-cloud-core==0.25.0
google-cloud-language==0.25.0
google-cloud-translate==0.25.0
google-gax==0.15.13
googleapis-common-protos==1.5.2
grpcio==1.4.0
httplib2==0.10.3
idna==2.5
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauth2client==3.0.0
ply==3.8
proto-google-cloud-language-v1beta2==0.15.3
protobuf==3.3.0
pyasn1==0.3.1
pyasn1-modules==0.0.10
pylint==1.7.2
requests==2.18.2
rsa==3.4.2
six==1.10.0
urllib3==1.22
wrapt==1.10.10
